:style :art: 완료하기 버튼 위치 이동
